Make Dockable track its parent DockSpace
This wont clear the property when a DockSpace is removed from the UI since that is a rabbit hole I dont really want to get into right now... but even so I dont think that is strictly needed. If a DockSpace is removed from the UI you can't access it from the Bento instance so both the Dockable and DockSpace should be GC'd at some later time.